Fixed bug with splitting in multi-stream decoder.
The split may leave the left piece of data look like unfinished utf-8.
This happens if some continuation bytes are missing just before the split.
E.g. in this case last 4-byte code point is unfinished:
  ... [f0 a7 a7?  |||  [f0 98 a6 94] [29] [20] ...

The fix is to always check not-last part to be fully processed.
If it is not, then it is treated as conversion error.